requirements.txt
flask
python-dotenv
google-generativeai
assemblyai
pytube
PyPDF2
streamlit
pathlib
requests
youtube_transcript_api